Fix bimanual task launch and update notebook for both tasks
- Fix 4 bimanual spec bugs preventing env startup:
  1. Delete extra keyframes (keys[1:]) before attach to avoid repeated
     name 'bimanual/' error during spec.compile()
  2. Strip material="prosthesis/..." from <site> elements — spec.attach()
     prefixes material definitions but not site references, causing
     compile-time "material not found" errors
  3. Remove unnamed TOUCH sensor so scene.py doesn't call mj_model.sensor("")
  4. Cast key_qpos/ctrl_range tensors to float32 to match sim dtype
- Update notebooks/train.ipynb: add task selection dropdown for both
  DieReorient and Bimanual-2024 tasks; use TASK_ID throughout

+++ src/mjlab_myochallenge/mdp/bimanual/actions.py
@@ -8,10 +8,9 @@ and write ctrl values directly via the sim's data object.
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-import mujoco
 import torch
 
 from mjlab.managers.action_manager import ActionTerm, ActionTermCfg
@@ -57,8 +56,8 @@ class MplJointPositionAction(ActionTerm):
         # Store ctrl ranges for [-1, 1] → [lo, hi] mapping
         ctrl_ids_np = self._ctrl_ids.cpu().numpy()
         ctrl_range = env.sim.mj_model.actuator_ctrlrange[ctrl_ids_np]
-        self._ctrl_lo = torch.tensor(ctrl_range[:, 0], device=env.device)
-        self._ctrl_hi = torch.tensor(ctrl_range[:, 1], device=env.device)
+        self._ctrl_lo = torch.tensor(ctrl_range[:, 0], device=env.device, dtype=torch.float32)
+        self._ctrl_hi = torch.tensor(ctrl_range[:, 1], device=env.device, dtype=torch.float32)
         self._ctrl_mid = 0.5 * (self._ctrl_lo + self._ctrl_hi)
         self._ctrl_half = 0.5 * (self._ctrl_hi - self._ctrl_lo)
